+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added additional options to `SearchTextField` to allow for custom icons and colors. Improved overall look and spacing of the search text field.
 * Moved `shapeBorder` from `DropDown` class to `DropDownOptions` as `border`. Changed default to a rounded border of `24.0` instead of `15.0`.
 * Added new constructors and static helper methods to `SelectedListItem` class.
+* Added option to autofocus `SearchTextField`.
 
 ## 0.0.1